core/state/snapshot: move iterator out into its own files
